pyproject.toml
[tool.poetry]
name = "sunbot"
version = "0.1.0"
description = ""
authors = ["Blake Hildebrand <baugusthildebrand@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.1.0"
requests = "^2.28.2"
psycopg2 = "^2.9.5"

[tool.poetry.scripts]
sunbot = "sunbot.sunbot:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
